refactor(FR-3425): move the errors-only sub-step rule to a neutral module
Quality pass on the fix. No behavior change.

- `isNonSuccessSubStep` and `SchedulingHistoryExpandableSubStep` move from the
  hook to `helper/schedulingHistory.ts`. Having `BAISubStepNodes` import a plain
  predicate out of a hook module inverted the layering: a hook used by the three
  table orchestrators became a dependency of a leaf component below them. Both
  sides now import from a neutral module that owns the rule they must agree on.
- `dataSignature` folds sub-step state in as two booleans rather than a raw
  count plus a boolean — the count contradicted the reason the boolean was there,
  since a sub-step arriving would re-apply the master mode and discard the user's
  manual row toggles even though expandability had not changed.
- The e2e spec's kebab-menu mode switch, repeated at four call sites, becomes a
  `switchExpandMode` helper alongside the file's existing helpers.
