+++ packages/backend.ai-ui/src/helper/index.ts
@@ -476,14 +476,11 @@ export const convertToUUID = (id: string): string => {
 };
 
 export * from './newLineToBrElement';
+export * from './schedulingHistory';
 export * from './useDebouncedDeferredValue';
 
 export type SemanticColor =
-  | 'success'
-  | 'info'
-  | 'warning'
-  | 'error'
-  | 'default';
+  'success' | 'info' | 'warning' | 'error' | 'default';
 
 export const useSemanticColorMap = (): Record<SemanticColor, string> => {
   const { token } = theme.useToken();
